application: Fix almost all compiler warning reported by clang for rv32/rv64

+++ application/baremetal/benchmark/dhrystone/dhry_1.c
@@ -335,7 +335,7 @@ int main(void)
     printf("Arr_2_Glob[8][7]:    %d\n", Arr_2_Glob[8][7]);
     printf("        should be:   Number_Of_Runs + 10\n");
     printf("Ptr_Glob->\n");
-    printf("  Ptr_Comp:          %d\n", (int) Ptr_Glob->Ptr_Comp);
+    printf("  Ptr_Comp:          %ld\n", (long) Ptr_Glob->Ptr_Comp);
     printf("        should be:   (implementation-dependent)\n");
     printf("  Discr:             %d\n", Ptr_Glob->Discr);
     printf("        should be:   %d\n", 0);
@@ -346,7 +346,7 @@ int main(void)
     printf("  Str_Comp:          %s\n", Ptr_Glob->variant.var_1.Str_Comp);
     printf("        should be:   DHRYSTONE PROGRAM, SOME STRING\n");
     printf("Next_Ptr_Glob->\n");
-    printf("  Ptr_Comp:          %d\n", (int) Next_Ptr_Glob->Ptr_Comp);
+    printf("  Ptr_Comp:          %ld\n", (long) Next_Ptr_Glob->Ptr_Comp);
     printf("        should be:   (implementation-dependent), same as above\n");
     printf("  Discr:             %d\n", Next_Ptr_Glob->Discr);
     printf("        should be:   %d\n", 0);
